incorporate queue processor class
