+++ openc-companies.py
@@ -16,7 +16,121 @@ import gzip
 import io
 import hashlib
 
-from IOQueueProcessor import IOQueueProcessor
+import multiprocessing
+from queue import Empty, Full
+
+class IOQueueProcessor():
+
+    def __init__(self, input_class, output_class, **kwargs):
+
+        self.process_count = kwargs.get('process_count', multiprocessing.cpu_count())
+        self.all_stop = multiprocessing.Value('i', 0)
+
+        self.input_class = input_class
+        self.output_class = output_class
+        self.input_queue = multiprocessing.Queue(self.process_count * 10)
+        self.output_queue = multiprocessing.Queue(self.process_count * 10)
+
+        self.input_queue_read_cnt = multiprocessing.Value('i', 0)
+        self.output_queue_read_cnt = multiprocessing.Value('i', 0)
+
+        self.kwargs = kwargs
+        self.process_list = []
+
+    def start_up(self):
+
+        self.process_list.append(multiprocessing.Process(target=self.output_queue_reader, args=(0, self.output_queue, self.output_class), kwargs=self.kwargs))
+        for process_number in range(self.process_count - 1):
+            self.process_list.append(multiprocessing.Process(target=self.input_queue_reader, args=(process_number+1, self.input_queue, self.output_queue, self.input_class), kwargs=self.kwargs))
+        for process in self.process_list:
+            process.start()
+
+    def finish_up(self):
+
+        # wait for queues
+        try:
+            while self.input_queue.qsize() or self.output_queue.qsize():
+                print(f"waiting for {self.input_queue.qsize()} input and {self.output_queue.qsize()} output queue records")
+                time.sleep(1)
+        except: # qsize does not work on mac
+            while not self.input_queue.empty() or not self.output_queue.empty():
+                if not self.input_queue.empty():
+                    print("waiting for input queue to finish")
+                elif not self.output_queue.empty():
+                    print("waiting for output queue to finish")
+                time.sleep(1)
+
+        with self.all_stop.get_lock():
+            self.all_stop.value = 1
+
+        start = time.time()
+        while time.time() - start <= 15:
+            if not any(process.is_alive() for process in self.process_list):
+                break
+            time.sleep(1)
+
+        for process in self.process_list:
+            if process.is_alive():
+                print(process.name, 'did not terminate gracefully')
+                process.terminate()
+            process.join()
+
+        self.input_queue.close()
+        self.output_queue.close()
+
+    def queue_read(self, q):
+        try:
+            return q.get(True, 1)
+        except Empty:
+            return None
+
+    def queue_write(self, q, msg):
+        while True:
+            try:
+                q.put(msg, True, 1)
+            except Full:
+                continue
+            break
+
+    def input_queue_reader(self, process_number, input_queue, output_queue, function_ref, **kwargs):
+
+        kwargs['process_number'] = process_number
+        input_class = function_ref(**kwargs)
+
+        while self.all_stop.value == 0:
+            queue_data = self.queue_read(input_queue)
+            if queue_data:
+                with self.input_queue_read_cnt.get_lock():
+                    self.input_queue_read_cnt.value += 1
+                result = input_class.run(queue_data)
+                if result:
+                    self.queue_write(output_queue, result)
+
+        input_class.close()
+
+    def output_queue_reader(self, process_number, output_queue, function_ref, **kwargs):
+
+        kwargs['process_number'] = process_number
+        output_class = function_ref(**kwargs)
+
+        while self.all_stop.value == 0:
+            queue_data = self.queue_read(output_queue)
+            if queue_data:
+                with self.output_queue_read_cnt.get_lock():
+                    self.output_queue_read_cnt.value += 1
+                output_class.run(queue_data)
+
+        output_class.close()
+
+    def process(self, msg):
+        self.queue_write(self.input_queue, msg)
+
+    def get_input_queue_read_cnt(self):
+        return self.input_queue_read_cnt.value
+
+    def get_output_queue_read_cnt(self):
+        return self.output_queue_read_cnt.value
+
 
 class writer():
